Console can display results in columns; Tab completion for commands Tab completion for get/setVar Have multiple backing textures - switch by name Console backing texture, num columns and font size from config file

+++ hdr/classes/paraFont.h
@@ -15,17 +15,17 @@ public:
 
 	void setColor(Uint8 red, Uint8 green, Uint8 blue, Uint8 alpha);
 
-	PARA_Surface *write(int X, int Y, const std::string &fontText);
+	PARA_Surface *write(float X, float Y, const std::string &fontText);
 
-	int lineHeight = 0;
+	float    lineHeight = 0;
 	SDL_Rect pos;
 
 private:
-	bool                fontSystemAvailable = false;
-	bool                available           = false;
-	TTF_Font            *fontHandle = nullptr;
-	PARA_Surface        *surface = nullptr;
-	PARA_Color          color = { 0,0,0,0 };
+	bool     fontSystemAvailable = false;
+	bool     available           = false;
+	TTF_Font *fontHandle         = nullptr;
+	PARA_Surface *surface        = nullptr;
+	PARA_Color   color           = {0, 0, 0, 0};
 };
 
 #endif //PARA_PARAFONT_H
